# Remove work-around for go 1.6 performance issue (forcing Windows timer to 1ms), improve battery life