@msftclas 2018-01-05
[![CLA assistant check](https://cla.opensource.microsoft.com/pull/badge/not_signed)](https://cla.opensource.microsoft.com/Microsoft/go-winio?pullRequest=68) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla.opensource.microsoft.com/Microsoft/go-winio?pullRequest=68) before we can accept your contribution.<br/><br/>:x: justfalter [sign now](https://cla.opensource.microsoft.com/Microsoft/go-winio?pullRequest=68)<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla.opensource.microsoft.com/check/Microsoft/go-winio?pullRequest=68) it.</sub>